[WIP][FEATURE] Localize the UserInterface
Implemented Ember-i18n in order to localize the Handlebar templates
and Javascript strings. Translations can be put into JSON files and
two new settings are added:
- defaultLocale: the default locale for the backend interface
- localeIncludePaths: the includepaths for the json translation files
as used bu the javascript

For the Handlebars template, the helper works similar like the Fluid
ViewHelper, so it parses the string based on the key, or renders the
key if the key is not found::

	{{translate id="key"}}

In case no translation is found for the key, 'default content' is returned
as output. In case a translation is found for the key, that is returned::

	{{translate id="key" default="default content"}}

Alternative syntax::

	{{#translate id="key"}}content{{/translate}}

Possible arguments to `translate`:

- `id`, the identifier (required)
- `default`, default content
- `package`, packageKey (defaults to TYPO3.Neos)
- `source`, sourceCatalogue (defaults to Main)

[!ToBeRemoved!] The CommandController to create json files from xliff files can be used
in the following way::

  ./flow translate:createjsonfromxliff

It can be used to create a single file or all json files.

Change-Id: I87e34e603027682112b53874c050b9dd16e52749
Releases: master
Related: NEOS-61, NEOS-435